requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
matplotlib==3.5.1
numpy==1.23.5
pandas==1.4.3
Pillow==9.3.0
requests==2.28.1
scikit_learn==1.2.0
torch==2.0.1+cu118
torchsummary==1.5.1
torchvision==0.15.2+cu118
tqdm==4.62.3
phe==1.5.0
networkx==3.1
ipykernel==6.23.2
openpyxl==3.1.2
fastapi==0.103
uvicorn==0.27.0.post1
sqlmodel==0.0.14
transformers==4.39.0
PyYAML==6.0.1
tensorflow==2.12
grpcio==1.49.1
grpcio-tools==1.49.1
datasets==2.16.1
huggingface-hub==0.20.3
PyMySQL==1.1.0
python-multipart==0.0.7
mysqlclient==2.2.1
jinja2
loguru
peft
sentencepiece
nltk